requirements.txt
numpy~=1.16.2
pydicom~=1.3.0
flask~=1.1.2
pillow~=6.1.0
click~=7.0
werkzeug~=0.15.6